requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
mysql-connector-python==8.0.34
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2
pytest==7.4.2
gunicorn==21.2.0
